Removed drop-down menu from navbar
