+++ src/App.css
@@ -10,7 +10,7 @@ body {
 
 
 .navbar-c {
-    opacity: 0.7;
+    opacity: 0.9;
 }
 
 .search-bar {
